add unit tests and refactor components
