+++ src/components/presentational/StartArea.tsx
@@ -1,30 +1,30 @@
 import styled from "styled-components";
 
 interface StastsAreaProps {
-    repos: string;
-    followers: string;
-    following: string;
+  repos: string;
+  followers: string;
+  following: string;
 }
 
 export const StartArea = ({ repos, followers, following }: StastsAreaProps) => {
-    return (
-        <Container>
-            <Data>
-                <span>Repos</span>
-                <strong>{repos}</strong>
-            </Data>
+  return (
+    <Container>
+      <Data>
+        <span>Repos</span>
+        <strong>{repos}</strong>
+      </Data>
 
-            <Data>
-                <span>Followers</span>
-                <strong>{followers}</strong>
-            </Data>
+      <Data>
+        <span>Followers</span>
+        <strong>{followers}</strong>
+      </Data>
 
-            <Data>
-                <span>Following</span>
-                <strong>{following}</strong>
-            </Data>
-        </Container>
-    );
+      <Data>
+        <span>Following</span>
+        <strong data-testid="following">{following}</strong>
+      </Data>
+    </Container>
+  );
 };
 
 const Container = styled.div`
