+++ src/main.tsx
@@ -24,11 +24,6 @@ const GlobalStyle = createGlobalStyle`
   }
 
 `;
-// if (process.env.NODE_ENV === 'development') {
-//   const { worker } = require('./mocks/browser')
-//   worker.start()
-// }
-
 ReactDOM.createRoot(document.getElementById("root") as HTMLElement).render(
   <React.StrictMode>
     <GlobalStyle />
